// design/source/mebel-next-en/libs/enums/product.enum.ts
export enum ProductType {
	SOFA = 'SOFA',
	CORNER_SOFA = 'CORNER_SOFA',
	ARMCHAIR = 'ARMCHAIR',
	BED = 'BED',
	POUF = 'POUF',
	MATTRESS = 'MATTRESS',
	KIDS = 'KIDS',
}

export enum ProductStatus {
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}

export enum ProductMaterial {
	FABRIC = 'FABRIC',
	VELVET = 'VELVET',
	MICROFIBER = 'MICROFIBER',
	LEATHER = 'LEATHER',
	ECO_LEATHER = 'ECO_LEATHER',
}

export enum ProductLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
}